rewrite: Move parameter rewrite to package (#499)
